chore: move test to separate class
